pyproject.toml
[tool.poetry]
name = "open_elections_peru"
version = "0.1.0"
description = "TODO: Write project short description."
authors = ["Reiner Marquez <rmaceissoft@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.1.2"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
django-model-utils = "^4.1.1"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django_redis_cache = "^3.0.0"
djangorestframework = "^3.12.2"
python-jne-peru = {git = "https://github.com/rmaceissoft/python-jne-peru"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
